requirements.txt
Django==1.11.12
cvxopt==1.1.9
cvxpy==0.4.11
mysqlclient==1.3.12
mysql-connector-python==2.1.7
numpy==1.14.2
pip==9.0.1
pylint==1.8.4
pytz==2018.4
scipy==1.0.1
wheel==0.31.0
djangorestframework==3.8.2

